requirements.txt
# Microscopy Image Analyzer Requirements

# Core dependencies
Pillow>=8.0.0
numpy>=1.20.0
imageio>=2.9.0

# Optional dependencies for enhanced functionality
# matplotlib>=3.3.0  # For advanced plotting features (future enhancement)
# scikit-image>=0.18.0  # For advanced image processing (future enhancement)
# opencv-python>=4.5.0  # For computer vision features (future enhancement)

# Development dependencies (uncomment for development)
# pytest>=6.0.0
# pytest-cov>=2.10.0
# black>=21.0.0
# flake8>=3.8.0
